approval-engine/requirements.txt
oracledb>=2.0
pandas>=2.0
numpy>=1.26
statsmodels>=0.14
scikit-learn>=1.4
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0
scipy>=1.13
joblib>=1.3
